code/backend/requirements.txt
flask
apiflask
Flask-SQLAlchemy
mysql-connector-python 
pymysql
boto3
python-dotenv
gunicorn
Pillow
pyjwt
marshmallow_sqlalchemy
Flask-JWT-Extended
Flask-Migrate
cryptography